collaborators fully removed before board deletion, created Board object
Collaborators are now fully removed before a board is deleted. There are
now longer stray "Link" nodes in the database. While doing this, I
realised boards on the Boards page are not displayed in the correct
order. To solve this, I changed the cypher query. I also renamed the
Board class in the backend to BoardTask, and created a new object called
Board, that is a model for what the Boards.GetAsync() function returns. This commit got a bit
bigger than it should have, with a bit too much going on at once.